Added platform layer for Windows and changed files.

+++ engine/src/core/asserts.h
@@ -16,7 +16,7 @@
 
 KEY_API void report_assertion_failure(const char *expression,
                                       const char *message,
-                                      const char *file, i32 line);
+                                      const char *file, i64 line);
 
 #define KASSERT(expression)                                                \
     {                                                                      \
